feat: implement FAQs component for new students
